edited the sales ui
