+++ src/app/dashboard/sales/new/page.tsx
@@ -1,8 +1,8 @@
 // "use client";
-import Sales from "~/components/accounting/Sales";
+import Sales from "~/app/dashboard/sales/_components/Sales";
 export default function Page() {
   return (
-    <main className="ml-5 mt-10  md:ml-10 md:w-[600px]">
+    <main className="ml-5 pt-10  md:ml-[70px] md:w-[600px]">
       <Sales />
     </main>
   );
