+++ src/components/ui/LoadingSkeleton.tsx
@@ -1,8 +1,14 @@
+import React from "react";
+import { cn } from "~/utils/utils";
+interface SpinnerProps extends React.HTMLProps<HTMLDivElement> {}
+
 export function Spinner() {
   return (
     <svg
       aria-hidden="true"
-      className="fill-primary h-8 w-8 animate-spin text-gray-200 dark:text-gray-600"
+      className={cn(
+        `h-8 w-8 animate-spin fill-primary text-gray-200 dark:text-gray-600`,
+      )}
       viewBox="0 0 100 101"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
